# docs/api/openapi.yaml
openapi: '3.1.1'
info:
  title: A365 Gateway API
  description: 'Connect independently hosted agents to Microsoft Entra, Agent 365, Purview and Prompt Shields. Administration uses an Entra access token; agent ingress uses a gateway-issued secret. Both use Authorization: Bearer, but are not interchangeable. Evaluate → call your model only on allow → submit the interaction. HTTP 202 means queued, never confirmed delivery. All timestamps are UTC. Interactive requests act on the live deployment.'
  version: v1
servers:
  - url: http://127.0.0.1:5080/
paths:
...
  /api/v1/agent-runtime/readiness:
    get:
      tags:
        - AgentRuntime
      summary: Check agent credential access
      description: "Returns 204 when this agent credential passes authentication and admission. This is not proof of Purview enforcement or telemetry delivery.\n\nAuthorization policy: ExternalAgentOnly."
      operationId: AgentRuntime.GetReadiness
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - AgentKey: [ ]
...
  /api/v1/protection/purview/policies:
    get:
      tags:
        - PurviewPolicyCatalog
      summary: Read existing Purview policies
      description: "Returns a fresh tenant-bound catalog and compatibility reasons. Failure returns 503 instead of a fabricated empty catalog. Policy definitions and classifiers are managed in Purview.\n\nAuthorization policy: AdministratorOnly."
      operationId: PurviewPolicyCatalog.List
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PurviewPolicyCatalogResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/PurviewPolicyCatalogResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/PurviewPolicyCatalogResponse'
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
  /api/v1/system/config:
    get:
      tags:
        - System
      summary: Read gateway settings
      description: "Returns supported runtime defaults and their current rowVersion. Initial tenant connection setup belongs to bootstrap.\n\nAuthorization policy: AdministratorOnly."
      operationId: System.GetSystemConfig
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SystemConfigDto'
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfigDto'
            text/json:
              schema:
                $ref: '#/components/schemas/SystemConfigDto'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
    patch:
      tags:
        - System
      summary: Update gateway defaults
      description: "Protection default changes require Idempotency-Key and If-Match headers matching the reviewed request body. Explicit per-agent protection choices remain unchanged.\n\nAuthorization policy: AdministratorOnly."
      operationId: System.UpdateSystemConfig
      parameters:
        - name: Idempotency-Key
          in: header
          description: Required for protection changes; must match body idempotencyKey (UUID v4).
          schema:
            type: string
        - name: If-Match
          in: header
          description: Required for protection changes; one exact row version matching body expectedRowVersion. Read the latest resource before reviewing a new change.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSystemConfigRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateSystemConfigRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateSystemConfigRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SystemConfigDto'
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfigDto'
            text/json:
              schema:
                $ref: '#/components/schemas/SystemConfigDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
components:
  schemas:
...
    AgentRegistrationDefaultsDto:
      type: object
      properties:
        environment:
          type: string
          default: Production
        reason:
          type:
            - 'null'
            - string
...
    PolicyCompatibilityResponse:
      required:
        - canAssign
        - reason
      type: object
      properties:
        canAssign:
          type: boolean
        reason:
          type:
            - 'null'
            - string
    ProblemDetails:
      type: object
      properties:
        type:
          type:
            - 'null'
            - string
        title:
          type:
            - 'null'
            - string
        status:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - 'null'
            - integer
            - string
          format: int32
        detail:
          type:
            - 'null'
            - string
        instance:
          type:
            - 'null'
            - string
...
    PurviewPolicyCatalogResponse:
      required:
        - tenantId
        - retrievedAtUtc
        - source
        - items
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        retrievedAtUtc:
          type: string
          format: date-time
        source:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurviewPolicyResponse'
    PurviewPolicyResponse:
      required:
        - id
        - displayName
        - mode
        - enforcementPlanes
        - individualApplicationIds
        - revision
        - compatibility
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        mode:
          type: string
        enforcementPlanes:
          type: array
          items:
            type: string
        individualApplicationIds:
          type: array
          items:
            type: string
            format: uuid
        revision:
          type: string
        compatibility:
          $ref: '#/components/schemas/PolicyCompatibilityResponse'
...
    SystemConfigDto:
      required:
        - provisioningMode
        - defaultObservabilityMode
        - retentionDaysIdempotencyRecords
        - rateLimitPerClient
        - rateLimitPerAgent
        - rateLimitGlobal
      type: object
      properties:
        provisioningMode:
          type: string
        defaultObservabilityMode:
          type: string
        retentionDaysIdempotencyRecords:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - integer
            - string
          format: int32
        rateLimitPerClient:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - integer
            - string
          format: int32
        rateLimitPerAgent:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - integer
            - string
          format: int32
        rateLimitGlobal:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - integer
            - string
          format: int32
        defaultAgent365ObservabilityEnabled:
          type:
            - 'null'
            - boolean
        defaultAzureMonitorExportEnabled:
          type:
            - 'null'
            - boolean
        provisioningExecutionEnabled:
          type: boolean
          default: false
        defaultPromptShieldEnabled:
          type: boolean
          default: false
        promptShieldAvailable:
          type: boolean
          default: false
        rowVersion:
          type:
            - 'null'
            - string
        registrationDefaults:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/AgentRegistrationDefaultsDto'
...
    UpdateSystemConfigRequest:
      required:
        - defaultObservabilityMode
        - retentionDaysIdempotencyRecords
        - rateLimitPerClient
        - rateLimitPerAgent
        - rateLimitGlobal
      type: object
      properties:
        defaultObservabilityMode:
          type:
            - 'null'
            - string
        retentionDaysIdempotencyRecords:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - 'null'
            - integer
            - string
          format: int32
        rateLimitPerClient:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - 'null'
            - integer
            - string
          format: int32
        rateLimitPerAgent:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - 'null'
            - integer
            - string
          format: int32
        rateLimitGlobal:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - 'null'
            - integer
            - string
          format: int32
        defaultAgent365ObservabilityEnabled:
          type:
            - 'null'
            - boolean
        defaultAzureMonitorExportEnabled:
          type:
            - 'null'
            - boolean
        defaultPromptShieldEnabled:
          type:
            - 'null'
            - boolean
        idempotencyKey:
          type:
            - 'null'
            - string
          format: uuid
        expectedRowVersion:
          type:
            - 'null'
            - string
...
  securitySchemes:
    EntraBearer:
      type: http
      description: Entra access token for the Gateway API audience with the endpoint's required gateway role. Do not paste a gateway agent key here.
      scheme: bearer
      bearerFormat: JWT
    AgentKey:
      type: http
      description: One-time gateway-issued agent key (a365gw_v1_...). Scoped to its registered agent. Do not use an Entra token for agent ingress.
      scheme: bearer
...