# src/Gateway.Api/Controllers/HealthController.cs
using Gateway.Infrastructure.Persistence;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace Gateway.Api.Controllers;

[ApiController]
[Route("")]
[AllowAnonymous]
public class HealthController : ControllerBase
{
    [HttpGet("health")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public IActionResult GetHealth()
    {
        return Ok(new { status = "Healthy" });
    }

    [HttpGet("health/ready")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status503ServiceUnavailable)]
    public async Task<IActionResult> GetReadiness(
        [FromServices] IDatabaseHealthProbe database,
        CancellationToken cancellationToken)
    {
        try
        {
            if (!await database.CanConnectAsync(cancellationToken))
            {
                return StatusCode(
                    StatusCodes.Status503ServiceUnavailable,
                    new { status = "Unavailable" });
            }
            return Ok(new { status = "Ready" });
        }
        catch
        {
            return StatusCode(
                StatusCodes.Status503ServiceUnavailable,
                new { status = "Unavailable" });
        }
    }

}